requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
idna==2.8
mysqlclient==1.4.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
